requirements.txt
# external requirements
paho-mqtt==2.0.0
pymysql==1.1.0
python-dotenv>=0.5.1
sqlalchemy==2.0.27
